Interact with PLUS.nl via Headless Browser

Provides functions to log in, add products to the cart, retrieve the current cart, and navigate the PLUS.nl website using a Chromote browser session.

plus_login(
  credentials = getOption("plus_credentials", default =
    system.file("login_credentials.yaml", package = "shinyplus")),
  info = interactive()
)

plus_logout(info = interactive())

plus_add_products(x, quantity = 1, info = interactive(), ...)

plus_current_cart(..., info = interactive())

plus_checkout()

plus_current_url(b = plus_env$browser)

plus_open_browser()

plus_page_reload()

Arguments

credentials

Path to a YAML file containing fields email and password, or a list contains those names. Can be set with options(plus_credentials = "...")

info

Logical to print info, default is TRUE in interactive sessions.

x

Name of the product, such as "PLUS Houdbare Halfvolle Melk Pak 1000 ml", or URL of the product, such as "plus-houdbare-halfvolle-melk-pak-1000-ml-957806".

quantity

Number of items to add to cart.

...

arguments passed to plus_login().

b

Browser to use.

Disclaimer

This open-source R package, shinyplus, is an independent software project and is not affiliated with, endorsed by, or associated in any way with PLUS Retail B.V. or the website plus.nl.

The package does not contain or store any data from plus.nl. It contains no product information, pricing data, or content originating from PLUS Retail B.V. or its systems. Users are required to provide their own data (e.g., product lists) manually.

All interactions with plus.nl occur through local, client-side browser automation, using the chromote package. This mechanism emulates a user operating a private browser session on their own machine. For example, when a user submits a basket to their PLUS shopping cart, this package simply opens a browser in the background and automates user-visible steps — such as logging in, navigating to product pages, and clicking buttons — exactly as a person would. No API calls or server-side data access is used or implied.

Any logos, icons, or brand references (e.g., the PLUS logo) used in the user interface are publicly available media assets sourced from Wikipedia under appropriate free-use licensing. All trademarks, service marks, and product names are the property of their respective owners.

Examples

if (FALSE) { # \dontrun{
plus_login()

plus_add_products("Houdbare halfvolle melk pak 1000 ml")
} # }